refactor: move vector to contrib
