+++ tests/test_contrib.py
@@ -1,0 +1,18 @@
+import json
+
+import pgcopy.contrib.vector
+
+from .test_datatypes import TypeMixin
+
+
+class TestVector(TypeMixin):
+    copy_manager_class = pgcopy.contrib.vector.CopyManager
+    extensions = ["vector"]
+    extension_types = ["vector"]
+    datatypes = ["vector"]
+    data = [
+        ((-1.5, 0, 2.3),),
+    ]
+
+    def cast(self, v):
+        return tuple(json.loads(v))
